Create FAST test base class
